requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0

# HTTP client and data validation
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# AI/LLM integration
google-generativeai==0.3.2

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Web scraping dependencies
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
playwright==1.40.0

# Security and encryption
cryptography==41.0.7

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
aiofiles==23.2.1

# Frontend (Streamlit)
streamlit==1.28.0
requests==2.31.0